unify UpdaterMuvT (Implicit)
