+++ hoomd/RNGIdentifiers.h
@@ -33,6 +33,13 @@ struct RNGIdentifier
     static const uint32_t UpdaterClustersPairwise = 0x50060112;
     static const uint32_t UpdaterExternalFieldWall = 0xba015a6f;
     static const uint32_t UpdaterMuVT = 0x186df7ba;
+    static const uint32_t UpdaterMuVTDepletants1 = 0xbbaa6272;
+    static const uint32_t UpdaterMuVTDepletants2 = 0x27568efe;
+    static const uint32_t UpdaterMuVTDepletants3 = 0x97bafe02;
+    static const uint32_t UpdaterMuVTDepletants4 = 0xb8b8b8b8;
+    static const uint32_t UpdaterMuVTDepletants5 = 0x81bcbde0;
+    static const uint32_t UpdaterMuVTDepletants6 = 0x9123abef;
+    static const uint32_t UpdaterMuVTPoisson = 0x7377befa;
     static const uint32_t UpdaterMuVTBox1 = 0x05d4a502;
     static const uint32_t UpdaterMuVTBox2 = 0xa74201bd;
     static const uint32_t ActiveForceCompute = 0x7edf0a42;
